requirements.txt
Pillow==10.1.0
mss==9.0.1
